+++ src/tests/helpers/storage-manager-stub.js
@@ -5,9 +5,6 @@ import { Promise } from 'rsvp';
 
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 
-// TODO this service stub is probably faulty, because there were problems
-// using it in manage-cluster-storages-test
-
 export default Service.extend({
   /**
    * Add storages here in tests (key: storage id, value: storage object)
@@ -23,8 +20,8 @@ export default Service.extend({
     return PromiseObject.create({
       promise: new Promise(resolve => {
         let storageDetailsList = A();
-        for (let storage in this.get('__storages')) {
-          storageDetailsList.push(this.getStorageDetails(storage.id));
+        for (let storageId in this.get('__storages')) {
+          storageDetailsList.push(this.getStorageDetails(storageId));
         }
         resolve(ArrayProxy.create({ content: storageDetailsList }));
       }),